fix(dir-filter): properly detect upward and downwards for project files
This change updates the dir_filter to use vim.fs.relpath to detect if
the file is a decendant or parent of a relevant project from the
solution.

This change should fix the error as well as be more performant because
we do not need to scan the directory for every file checked.